Fixed LogoDisplay Added LogoDisplay to the Fixed the css for the drag-n-drop type Repaired the JS file Configured the JS File to hit breakpoints

+++ MentorBilling/Miscellaneous/Size.cs
@@ -7,7 +7,7 @@ namespace MentorBilling.Miscellaneous
 {
     public class Size
     {
-        public Int32 Width { get; set; }
-        public Int32 Height { get; set; }
+        public Double Width { get; set; }
+        public Double Height { get; set; }
     }
 }
